requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
pyttsx3==2.90
RPi.GPIO==0.7.1
cvlib==0.2.7
Pillow==10.0.0
matplotlib==3.7.2
ultralytics==8.0.196
pygame==2.5.2